refactor: move config and db sources to own dirs
