+++ src/github/client/token-storage.ts
@@ -2,7 +2,7 @@ import { RedisClient } from "redis";
 import { inject } from "inversify";
 import { injectable } from "inversify";
 import { LOGGER } from "../../logger";
-import RedisClientFactory, { DATABASE_INDEX } from "../../redis-client";
+import RedisClientFactory, { DATABASE_INDEX } from "../../db/redis-client";
 
 @injectable()
 class TokenStorage {
