+++ src/pages/page-routes.ts
@@ -4,8 +4,8 @@ import { Router, Request, Response, NextFunction } from "express";
 import * as express from "express";
 import { injectable } from "inversify";
 import { inject } from "inversify";
-import { ConfigurationService } from "../configuration";
-import RedisClientFactory from "../redis-client";
+import { ConfigurationService } from "../config/configuration";
+import RedisClientFactory from "../db/redis-client";
 
 @injectable()
 class PageRoutes {
